refactor(internal): Remove dependence of Promise.all helpers on AggregateError
Work around absence of AggregateError in xsnap.
