+++ packages/swing-store/src/snapStore.js
@@ -4,7 +4,7 @@ import { finished as finishedCallback, PassThrough, Readable } from 'stream';
 import { promisify } from 'util';
 import { createGzip, createGunzip } from 'zlib';
 import { Fail, q } from '@endo/errors';
-import { withDeferredCleanup } from '@agoric/internal/src/node/utils.js';
+import { withDeferredCleanup } from '@agoric/internal';
 import { buffer } from './util.js';
 
 /**
